backend/google-analytics/requirements.txt
google-analytics-data==0.18.4
google-auth==2.23.4
